release: 1.2.945b

+++ src/main/java/difficultymod/capabilities/temperature/ITemperature.java
@@ -8,7 +8,7 @@ public interface ITemperature extends IBaseCapability {
 	/**
 	 * Retrieve the temperature penalty as result of a player's armor.
 	 */
-	int GetPlayerArmorPenalty(EntityPlayer player);
+	float GetPlayerArmorPenalty(EntityPlayer player);
 
 	/**
 	 * Get the current temperature with modifiers applied.
